pyproject.toml
[tool.poetry]
name = "dataset-manager"
version = "0.2.0"
description = "Scalable dataset management platform with ETL and advanced features"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
cassandra-driver = "^3.29.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-multipart = "^0.0.6"
pandas = "^2.1.0"
pyarrow = "^13.0.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
cryptography = "^41.0.0"

# Release 2 Dependencies
kafka-python = "^2.0.2"
boto3 = "^1.28.0"
redis = "^5.0.0"
apache-airflow = "^2.7.0"
prometheus-client = "^0.18.0"
requests = "^2.31.0"

# Testing
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
httpx = "^0.25.0"

[tool.poetry.dev-dependencies]
black = "^23.0.0"
flake8 = "^6.0.0"
mypy = "^1.0.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
